refactor: move toggle to widgets
